Introduce new composition printer / implementation
This forks from the newly updated schema printer. However, the goal
of this printer is to print the FINAL composition, rather than the
schema for a single federated service.

By peeking at the existing annotations in the `extensions` on the
composed `GraphQLSchema`, the composition printer can gather
important metadata like type and field ownership which is then
annotated in the final SDL.